put the account menu in the header
