+++ v3/ui/src/components/accounts/DepositingNav.tsx
@@ -20,61 +20,13 @@ export function DepositingNav() {
 
   return (
     <Flex alignItems="center" mb="10">
-      <Menu>
-        <MenuButton size="sm" as={Button} variant="outline" rightIcon={<ChevronDownIcon />}>
-          {id ? `Account #${id}` : 'Create Account'}
-        </MenuButton>
-        <MenuList fontSize="xs" px="2" bg="black" border="1px solid rgba(255,255,255,0.33)">
-          {userAccounts.map((account) => {
-            const isCurrentAccount = id === account.toString();
-            const menuItem = (
-              <MenuItem
-                key={account}
-                _hover={{ bg: 'whiteAlpha.200' }}
-                _focus={{ bg: 'whiteAlpha.200' }}
-                _active={{ bg: 'whiteAlpha.200' }}
-              >
-                <Flex width="100%" alignItems="center">
-                  {isCurrentAccount && <CheckIcon marginRight={1} />}
-
-                  {account}
-                </Flex>
-              </MenuItem>
-            );
-
-            return isCurrentAccount ? (
-              menuItem
-            ) : (
-              <RouterLink key={account} to={`/accounts/${account}${routingSearchParams}`}>
-                {menuItem}
-              </RouterLink>
-            );
-          })}
-          <MenuItem
-            _hover={{ bg: 'whiteAlpha.200' }}
-            _focus={{ bg: 'whiteAlpha.200' }}
-            _active={{ bg: 'whiteAlpha.200' }}
-          >
-            <Link
-              as={RouterLink}
-              to={`/accounts/create${routingSearchParams}`}
-              _focus={{ boxShadow: 'none' }}
-              _hover={{ textDecoration: 'none' }}
-              fontWeight="semibold"
-            >
-              Create new account
-            </Link>
-          </MenuItem>
-        </MenuList>
-      </Menu>
-      {Boolean(innerPage) ? (
+      {innerPage ? (
         <Link
           as={RouterLink}
           to={generatePath('/accounts/:id/*', {
             id,
             '*': routingSearchParams,
           })}
-          ml="auto"
           fontSize="xs"
           fontWeight="normal"
           color="cyan.500"
